Display hourly weather icons on temperature chart
Generates pointers to 32 pc bitmaps which are then drawn at the x-axis tick marks except for the first and last one and the above the temperature line.
Tested with different HOURLY_GRAPH_MAX.
Still readable if precipitation overlaps, however a grayscale option would be truly amazing for this (is this feasible?).
Currently does not differentiate between moon present or absent, night time is always with moon.
Only tested with DISP_BW_V2, therefore only enabled for this display.