Display hourly icons above temperature line (#136)
* Display hourly weather icons on temperature chart

Generates pointers to 32 pc bitmaps which are then drawn at the x-axis tick marks except for the first and last one and the above the temperature line.
Tested with different HOURLY_GRAPH_MAX.
Still readable if precipitation overlaps, however a grayscale option would be truly amazing for this (is this feasible?).
Currently does not differentiate between moon present or absent, night time is always with moon.
Only tested with DISP_BW_V2, therefore only enabled for this display.

* Change to temperature line alignment

The temperature line is shifted by half an hour I think. Maybe there is a good reason for this that I missed. This change aligns the line correctly and adds one more index for the last hour. However, it limits the number of hours to 47.

* Change to sleep duration

For some reason, for me it would always refresh 1 minute before the full or half hour. This bugged me, also because of the corresponding x-axis limits. This might just be a problem on my end though

* Revert "Change to sleep duration"

This reverts commit 6c12a0065fff82e975763421849947a340adb984.

* Add additional sleep time compensation to avoid waking a second before intended interval

* cleaned up hourly icon code.

less duplicate code,
updated from c++11 to c++17 for better constexpr support
icons now always display above the temperature line, but still stay as
close as possible.

---------

Co-authored-by: Luke Marzen <98473324+lmarzen@users.noreply.github.com>